moved noisy receiver to mainActivity
